{{cookiecutter.project_slug}}/requirements.txt
amqp==5.0.6
asgiref==3.3.4
celery==5.1.0
chardet==4.0.0
Django==3.2.4
django-dotenv==1.4.2
django-filter==2.4.0
dj-inmemorystorage==2.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
kombu==5.1.0
model-bakery==1.3.2
psycopg2==2.9.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
